// src/components/Illustrations/enums.ts
export enum EPictureSizes {
  Small = 'small', // Размер для иконок 100x100
  Medium = 'medium', // Размер для иконок 200x200
  Wide = 'wide', // Размер для иконок 300x164
  Standard = 'standard', // Размер для иконок 300x200
  Tall = 'tall', // Размер для иконок 300x232
  ExtraLarge = 'extra-large' // Размер для иконок 460x340
}

export enum EPictureNames {
  Anonymous = 'anonymous',
  Block = 'block',
  Calendar = 'calendar',
  Catch = 'catch',
  Chemistry = 'chemistry',
  Code = 'code',
  Coffee = 'coffee',
  Communication = 'communication',
  Dart = 'dart',
  Doc = 'doc',
  End = 'end',
  Error = 'error',
  Error2 = 'error2',
  Error3 = 'error3',
  Error404 = 'error404',
  Give = 'give',
  Globus = 'globus',
  Graphics = 'graphics',
  Graphics2 = 'graphics2',
  Key = 'key',
  Les = 'les',
  Lunch = 'lunch',
  Meet = 'meet',
  NoImage = 'no-image',
  Notes = 'notes',
  PictureDefault = 'picture-default',
  PictureError = 'picture-error',
  PictureFactory = 'picture-factory',
  PictureUploadDoc = 'picture-upload-doc',
  PictureUploadFactory = 'picture-upload-factory',
  PictureUploadPic = 'picture-upload-pic',
  PictureUploadVideo = 'picture-upload-video',
  Padenie = 'padenie',
  Padenie2 = 'padenie2',
  PageNotFound = 'page-not-found',
  Pyramid = 'pyramid',
  Rate = 'rate',
  Relax = 'relax',
  Rost = 'rost',
  Search = 'search',
  Sleep = 'sleep',
  Time = 'time',
  Treugolnic = 'treugolnic',
  Ves = 'ves',
  Questions = 'questions',
  Zamok = 'zamok'
}